Revert "Simplified marketplace demo"
