--- admin/README.md
@@ -1,9 +1,0 @@
-# Admin
-
-Simple Express server representing admin functions.
-
-```
-cd admin
-npm install
-npm start
-```
