--- admin/index.js
@@ -1,13 +1,0 @@
-import express from 'express';
-import cors from 'cors';
-
-const app = express();
-app.use(cors());
-app.use(express.json());
-
-app.get('/health', (req, res) => {
-  res.json({ status: 'ok' });
-});
-
-const PORT = process.env.PORT || 5000;
-app.listen(PORT, () => console.log('Admin server on ' + PORT));
