--- client/README.md
@@ -1,9 +1,0 @@
-# Client
-
-Next.js 15 app with Tailwind CSS. Provides basic pages for login, register and gig management.
-
-```
-cd client
-npm install
-npm run dev
-```
