--- server/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "marketplace-server",
-  "version": "1.0.0",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "start": "node index.js"
-  },
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "cors": "^2.8.5",
-    "express": "^4.18.2",
-    "jsonwebtoken": "^9.0.1",
-    "socket.io": "^4.7.2"
-  }
-}
